+++ src/main/java/org/folio/search/configuration/properties/SearchConfigurationProperties.java
@@ -71,10 +71,14 @@ public class SearchConfigurationProperties {
     private IndexingDataFormat dataFormat;
 
     /**
-     * Scheduler delay for indexing subjects/contributors/classifications.
-     * */
-    private long instanceChildrenIndexDelayMs;
+     * Defines if module should process subjects/contributors/classifications/call-numbers in a background.
+     */
+    private Boolean instanceChildrenIndexEnabled;
 
+    /**
+     * Scheduler delay for indexing subjects/contributors/classifications.
+     */
+    private long instanceChildrenIndexDelayMs;
 
     /**
      * Batch size for background processing sub-resources (contributors, subjects, classifications, etc.).