+++ src/main/java/org/folio/search/service/reindex/jdbc/MergeInstanceRepository.java
@@ -102,7 +102,7 @@ public class MergeInstanceRepository extends MergeRangeRepository {
   public SubResourceResult fetchByTimestamp(String tenant, Timestamp timestamp) {
     var sql = SELECT_BY_UPDATED_QUERY.formatted(getSchemaName(tenant, context.getFolioModuleMetadata()));
     var records = jdbcTemplate.query(sql, instanceRowMapper(), timestamp);
-    var lastUpdateDate = records.isEmpty() ? null : records.getLast().get(LAST_UPDATED_DATE_FIELD);
+    var lastUpdateDate = records.isEmpty() ? null : records.get(records.size() - 1).get(LAST_UPDATED_DATE_FIELD);
     return new SubResourceResult(records, (Timestamp) lastUpdateDate);
   }