+++ src/test/java/org/folio/search/service/scheduled/ScheduledInstanceSubResourcesServiceTest.java
@@ -20,6 +20,7 @@ import org.folio.search.configuration.properties.SearchConfigurationProperties;
 import org.folio.search.model.types.ReindexEntityType;
 import org.folio.search.service.InstanceChildrenResourceService;
 import org.folio.search.service.ResourceService;
+import org.folio.search.service.ScheduledInstanceSubResourcesService;
 import org.folio.search.service.reindex.jdbc.ItemRepository;
 import org.folio.search.service.reindex.jdbc.MergeInstanceRepository;
 import org.folio.search.service.reindex.jdbc.SubResourceResult;
@@ -142,8 +143,8 @@ class ScheduledInstanceSubResourcesServiceTest {
       .when(executionService).executeSystemUserScoped(anyString(), any());
     when(subResourcesLockRepository.lockSubResource(any(), any())).thenReturn(Optional.of(timestamp));
     when(tenantRepository.fetchDataTenantIds()).thenReturn(tenantIds);
-    mockSubResourceResult(tenantIds.getFirst(), timestamp);
-    mockSubResourceResult(tenantIds.getLast(), timestamp);
+    mockSubResourceResult(tenantIds.get(0), timestamp);
+    mockSubResourceResult(tenantIds.get(tenantIds.size() - 1), timestamp);
 
     // Act
     service.persistChildren();